requirements.txt
torch
torchvision
numpy
opencv-python
pillow
streamlit
matplotlib
scikit-learn
tqdm